requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
SQLAlchemy>=2.0
pydantic>=2.6
email-validator>=2.1
bcrypt>=4.1
python-jose[cryptography]>=3.3
qrcode>=7.4
pillow>=10.2
google-generativeai>=0.7
python-dotenv>=1.0
